chore: update sign in
