+++ apps/journeys-admin/src/components/SignIn/FacebookButton/index.ts
@@ -1,0 +1,1 @@
+export { FacebookButton } from './FacebookButton'
