+++ apps/journeys-admin/src/components/SignIn/GoogleButton/index.ts
@@ -1,0 +1,1 @@
+export { GoogleButton } from './GoogleButton'
